change uint type

+++ contracts/helper/TestLibrary.sol
@@ -14,28 +14,28 @@ contract TestLibrary {
   }
 
   function quote(
-    uint256 amountA,
-    uint256 reserveA,
-    uint256 reserveB
-  ) public pure returns (uint256 amountB) {
+    uint112 amountA,
+    uint112 reserveA,
+    uint112 reserveB
+  ) public pure returns (uint112 amountB) {
     amountB = AMMLibrary.quote(amountA, reserveA, reserveB);
   }
 
   function getAmountOut(
-    uint256 amountIn,
-    uint256 reserveIn,
-    uint256 reserveOut,
-    uint256 fee
-  ) public pure returns (uint256 amountOut) {
+    uint112 amountIn,
+    uint112 reserveIn,
+    uint112 reserveOut,
+    uint112 fee
+  ) public pure returns (uint112 amountOut) {
     amountOut = AMMLibrary.getAmountOut(amountIn, reserveIn, reserveOut, fee);
   }
 
   function getAmountIn(
-    uint256 amountOut,
-    uint256 reserveIn,
-    uint256 reserveOut,
-    uint256 fee
-  ) public pure returns (uint256 amountIn) {
+    uint112 amountOut,
+    uint112 reserveIn,
+    uint112 reserveOut,
+    uint112 fee
+  ) public pure returns (uint112 amountIn) {
     amountIn = AMMLibrary.getAmountIn(amountOut, reserveIn, reserveOut, fee);
   }
